Administrador productos

+++ ecommerce/src/app/admin/admin.component.ts
@@ -5,6 +5,8 @@ import { Router } from '@angular/router';
 // import { Producto } from '../models/producto.model';
 import { Producto } from '../models/producto';
 import { ProductoService } from '../services/producto.service';
+import { AuthService } from '../services/auth.service';
+import { ViewportScroller } from '@angular/common';
 
 @Component({
   selector: 'app-admin',
@@ -13,25 +15,167 @@ import { ProductoService } from '../services/producto.service';
 })
 export class AdminComponent implements OnInit {
   listProductos: Producto[] = [];
-  mostrarMensaje: boolean | undefined;
-  
+  totalPages: number = 1;
+  currentPage: number = 1;
+  itemsPerPage: number = 20;
+  searchQuery: string = '';
+  categoriasUnicas: string[] = [];
+  tallasUnicas: string[] = [];
+  listProductosOriginal: Producto[] = [];
+  sugerenciasProductos: Producto[] = [];
+  sugerenciasEnTiempoReal: Producto[] = [];
+
+  private routesMap = new Map([
+    ['login', '/login'],
+    ['inicio de sesión', '/login'],
+    ['carrito', '/cart'],
+    ['agregar','/agregar'],
+    ['informacion','/info'],
+    ['editar-producto','/editar-producto'],
+
+  ]);
+
   constructor(
-    @Inject(ProductoService) private _productoService: ProductoService,
-    private toastr: ToastrService
+    
+    private _productoService: ProductoService,
+    private toastr: ToastrService,
+    private router: Router,
+    private viewportScroller: ViewportScroller,
+    public authService: AuthService 
   ) {}
 
 
+
   ngOnInit(): void {
-    this.obtenerProducto();
-    // Simulamos una llamada asíncrona para obtener los productos
-    setTimeout(() => {
-      // Supongamos que después de 5 segundos, aún no hay productos
-      if (this.listProductos.length === 0) {
-        this.mostrarMensaje = false; // Ocultamos el mensaje
-      }
-    }, 8000); // 5000 milisegundos = 5 segundos
+    this.obtenerProductos(this.currentPage, this.itemsPerPage);
+    // this.obtenerProductosUnicos();
+      // Verifica si el usuario ya está autenticado al cargar el componente
+    const token = localStorage.getItem('token');
+    this.authService.isAuthenticated = !!token;
   }
 
+  editarProducto(): void {
+    this.router.navigate(['/editar-producto']);
+  }
+
+  navigateToLogin(): void {
+    this.router.navigate(['/login']);
+  }
+  logout(): void {
+    this.authService.logout();
+    this.toastr.success('A salido de la sesion!', 'Sesion cerrada!');
+    // Opcionalmente, puedes redirigir a la página de inicio o login después del cierre de sesión
+    this.router.navigate(['/login']);
+  }
+
+  buscarProductos(): void {
+    const query = this.searchQuery.toLowerCase().trim();
+    this.sugerenciasProductos = []; // Reiniciar las sugerencias
+
+    // Verifica si la consulta coincide con una ruta
+    for (const [keyword, path] of this.routesMap.entries()) {
+      if (query.includes(keyword)) {
+        // Si la consulta contiene la palabra clave, navega a la ruta correspondiente
+        if (path.includes(':')) {
+          // Si la ruta tiene un parámetro, pide al usuario que lo ingrese
+          const param = prompt(
+            `Ingresa el valor del parámetro para ${keyword}`
+          );
+          if (param) {
+            this.router.navigateByUrl(path.replace(':id', param));
+          }
+        } else {
+          this.router.navigateByUrl(path);
+        }
+        return; 
+      }
+    }
+
+    if (query === '') {
+      // Si la consulta está vacía, restablecer los filtros
+    } else {
+      // Filtrar productos por nombre o categoría que coincidan con la consulta
+      const resultados = this.listProductosOriginal.filter((producto) => {
+        return (
+          producto.nombre.toLowerCase().includes(query) ||
+          producto.categoria.toLowerCase().includes(query)
+        );
+      });
+
+      if (resultados.length === 0) {
+        // No se encontraron resultados, redirigir a la página de error
+        this.router.navigate(['/error']);
+      } else {
+        // Mostrar los resultados en la lista
+        this.listProductos = resultados;
+
+        // Obtener sugerencias de productos relacionados
+        this.sugerenciasProductos = resultados.slice(0, 5); // Tomar los primeros 5 resultados
+
+        // Hacer scroll hacia abajo después de la búsqueda de productos
+        setTimeout(() => {
+          this.viewportScroller.scrollToPosition([0, 730]);
+        }, 10);
+      }
+    }
+  }
+
+  obtenerProductos(page: number, itemsPerPage: number): void {
+    this._productoService.getProductos(page, itemsPerPage).subscribe(
+      (data) => {
+        // console.log('Datos de la API:', data);
+        this.listProductos = data.products;
+        this.totalPages = data.totalPages;
+        this.currentPage = page;
+        this.categoriasUnicas = this.obtenerCategoriasUnicas(
+          this.listProductos
+        );
+        // this.obtenerProductosUnicos();
+        this.listProductosOriginal = this.listProductos.slice(); // Guardar copia de la lista original
+      },
+      (error) => {
+        console.log(error);
+      }
+    );
+  }
+
+  private obtenerCategoriasUnicas(productos: Producto[]): string[] {
+    const categoriasSet = new Set<string>();
+    productos.forEach((producto) => {
+      categoriasSet.add(producto.categoria);
+    });
+    return Array.from(categoriasSet);
+  }
+
+  changePage(page: number): void {
+    this.obtenerProductos(page, this.itemsPerPage);
+  }
+
+  get totalPagesArray(): number[] {
+    return Array.from({ length: this.totalPages }, (_, i) => i + 1);
+  }
+
+  filtrarPorCategoria(categoria: string): void {
+    this.listProductos = this.listProductosOriginal.filter(
+      (producto) => producto.categoria === categoria
+    );
+  }
+
+  navigateToProductDetails(productoNombre: string): void {
+    // Buscar el producto en listProductosOriginal por su nombre
+    const producto = this.listProductosOriginal.find(p => p.nombre === productoNombre);
+  
+    // Si se encuentra el producto, navegar a sus detalles
+    if (producto) {
+      this.router.navigate(['/product-detail', producto._id]);
+    }
+  }
+  actualizarSugerencias(query: string): void {
+    this.sugerenciasEnTiempoReal = this.listProductosOriginal.filter(producto => {
+      const nombre = producto.nombre ? producto.nombre.toLowerCase() : '';
+      return nombre.includes(query.toLowerCase());
+    });
+  }
 
   obtenerProducto() {
     this._productoService.getProducto().subscribe((data: Producto[]) => {
@@ -43,19 +187,17 @@ export class AdminComponent implements OnInit {
   }
 
   eliminarProducto(id: any) {
-    this._productoService.eliminarProducto(id).subscribe(
-      (data: any) => {
-        this.toastr.success('El producto fue eliminado con éxito', 'Producto Eliminado');
-        this.obtenerProducto();
-        window.location.reload();
-      },
-      (error: any) => {
-        console.log(error);
-      }
-    );
+    if (confirm('¿Estás seguro de que deseas eliminar este producto?')) {
+      this._productoService.eliminarProducto(id).subscribe(
+        (data: any) => {
+          this.toastr.success('El producto fue eliminado con éxito', 'Producto Eliminado');
+          this.obtenerProducto();
+          window.location.reload();
+        },
+        (error: any) => {
+          console.log(error);
+        }
+      );
+    }
   }
-};
-
-
-
- 
+}
